fuzz/pyproject.toml
[project]
name = "fuzz"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "boofuzz>=0.4.2",
    "pyserial>=3.5",
]
